Health Bar and the Hearts
Create a canvas and import the hearts into that canvas. For the script put hearts_0 for empty heart and hearts_1 for full heart. put like 5 of the hearts_1 under the canvas and select and add those to the script. Check the script to, might want to change OnTriggerEnter to on OnCollisionEnter and create a tag for the enemies so the code knows to remove a heart if player is hit.